requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
slack-sdk==3.31.0
python-dotenv==1.0.1
PyYAML==6.0.2

# Dev / test
pytest==8.3.3
pytest-cov==5.0.0
